refactor: move reco fetch hook to a reco folder
