+++ src/features/home/api/useHomeRecommendedHits.test.ts
@@ -7,9 +7,9 @@ import {
   getRecommendationParameters,
   useHomeRecommendedHits,
 } from 'features/home/api/useHomeRecommendedHits'
-import * as recommendedIdsAPI from 'features/home/api/useHomeRecommendedIdsMutation'
 import { RecommendationParametersFields } from 'libs/contentful'
 import { env } from 'libs/environment'
+import * as recommendedIdsAPI from 'libs/recommendation/useHomeRecommendedIdsMutation'
 import { useSubcategoryLabelMapping } from 'libs/subcategories/mappings'
 import { renderHook } from 'tests/utils'
