+++ src/features/home/api/useHomeRecommendedHits.ts
@@ -1,13 +1,13 @@
 import { useEffect, useState } from 'react'
 
 import { computeBeginningAndEndingDatetimes } from 'features/home/api/helpers/computeBeginningAndEndingDatetimes'
-import {
-  RecommendedIdsRequest,
-  useHomeRecommendedIdsMutation,
-} from 'features/home/api/useHomeRecommendedIdsMutation'
 import { RecommendationParametersFields } from 'libs/contentful'
 import { env } from 'libs/environment'
 import { GeoCoordinates } from 'libs/geolocation'
+import {
+  RecommendedIdsRequest,
+  useHomeRecommendedIdsMutation,
+} from 'libs/recommendation/useHomeRecommendedIdsMutation'
 import { SearchHit } from 'libs/search'
 import { getCategoriesFacetFilters } from 'libs/search/utils'
 import { useSubcategoryLabelMapping } from 'libs/subcategories/mappings'
